Adding the abstract scenario definition
The abstract scenario definition creates an abstract road network that is representative of the TFHRC road network in which the demo is executed. The road network is made up of two junctions (west/east intersections), 3 roads connecting the junctions (representing Innovation Drive and the CIA loop), and two routes. Routes are defined for the live, virtual, and constructive vehicles existing in the simulation. The vehicles follow these routes and exhibit similar behavior to what occurs in the demonstration, forming a platoon, entering a work-zone, and stopping at intersections. Unfortunately, there is not support for traffic signal controllers in the OpenScenario v2.0 framework, so signal-interaction behavior is not explicitly defined in the simulation, but vehicles stop and start as they would in the actual demonstration.